+++ octopusCats/src/test/scala/octopus/async/cats/CatsEffectIOIntegrationSpec.scala
@@ -3,6 +3,6 @@ package octopus.async.cats
 import cats.effect.IO
 import octopus.AsyncValidationSpec
 import octopus.async.cats.implicits._
-import octopus.async.cats.ToFuture._
+import octopus.async.cats.ToFutureImplicits._
 
 class CatsEffectIOIntegrationSpec extends AsyncValidationSpec[IO]
