+++ octopusCats/src/test/scala/octopus/async/cats/MonixTaskIntegrationSpec.scala
@@ -4,6 +4,6 @@ import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import octopus.AsyncValidationSpec
 import octopus.async.cats.implicits._
-import octopus.async.cats.ToFuture._
+import octopus.async.cats.ToFutureImplicits._
 
 class MonixTaskIntegrationSpec extends AsyncValidationSpec[Task]
