+++ octopusScalaz/.jvm/src/test/scala/octopus/async/scalaz/ScalazTaskIntegrationSpec.scala
@@ -2,7 +2,7 @@ package octopus.async.scalaz
 
 import octopus.AsyncValidationSpec
 import octopus.async.scalaz.instances._
-import octopus.async.scalaz.ToFuture._
+import octopus.async.scalaz.ToFutureImplicits._
 import scalaz.concurrent.Task
 
 class ScalazTaskIntegrationSpec extends AsyncValidationSpec[Task]
